requirements.txt
functions-framework==3.*
google-generativeai>=0.7.0
flask==2.*